README
This project is an Android application that implements the best practices of the object-oriented paradigm and uses the language richness of the Kotlin programming language. The application supports multilingualism, managing all application resources, and using custom styles. Additionally, the application has advanced graphical interface elements, such as animations and Material design components. The basic architecture of the application is implemented using standard architectural components, such as Activity components, implicit and explicit Intent objects, menus and dialogs, management of interface Fragments, as well as persistent storage and data management of the application using ContentProvider and ContentResolver components. Moreover, asynchronous communication with the server is implemented using Service components, Handler and Looper libraries, and BroadcastReceiver components for sending notifications. Finally, various levels of security are explored and implemented using application permissions defined in the manifest, RuntimePermission permissions, PendingIntent components, and system services.